requirements.txt
contourpy
cycler
dateutil
filelock
fonttools
fsspec
functorch
jinja2
kiwisolver
markupsafe
matplotlib
mpmath
networkx
numpy
opencv-python
packaging
pandas
Pillow
pyparsing
python-dateutil
pytz
scipy
setuptools
six
style
sympy
torch
torchaudio
torchvision
typing-extensions
tzdata
update
